+++ tests/testthat/test-hospital_capacity.R
@@ -37,8 +37,8 @@ test_that("Closures: hospital capacity and closure time", {
   )
 
   expect_lt(
-    x$response_data$closure_info$closure_time_start,
-    y$response_data$closure_info$closure_time_start
+    x$response_data$closure_info$closure_times_start,
+    y$response_data$closure_info$closure_times_start
   )
 
   # hospital capacity override from `daedalus()`
@@ -51,8 +51,8 @@ test_that("Closures: hospital capacity and closure time", {
   )
 
   expect_gt(
-    x$response_data$closure_info$closure_time_start,
-    y$response_data$closure_info$closure_time_start
+    x$response_data$closure_info$closure_times_start,
+    y$response_data$closure_info$closure_times_start
   )
 })